app/requirements.txt
flask
flask.sqlalchemy
alembic
python-dotenv
mysql-connector-python
mysqlclient
dependency-injector